refactor: split types/variant.go into decode and encode files
Move decoding logic to variant_decode.go and encoding logic to
variant_encode.go; split matching tests into variant_decode_test.go
and variant_encode_test.go.
